Renamed solr update classes to clarify roles
