+++ services-camel/src/test/resources/cdr-events-context.xml
@@ -13,16 +13,16 @@
 		http://camel.apache.org/schema/spring
 		http://camel.apache.org/schema/spring/camel-spring.xsd">
 	
-	<bean id="solrUpdateProcessor" class="org.mockito.Mockito" factory-method="mock">
-		<constructor-arg value="edu.unc.lib.cdr.SolrUpdateProcessor" />
+	<bean id="cdrEventToSolrUpdateProcessor" class="org.mockito.Mockito" factory-method="mock">
+		<constructor-arg value="edu.unc.lib.cdr.CdrEventToSolrUpdateProcessor" />
 	</bean>
 	
 	<bean id="cdrEventProcessor" class="org.mockito.Mockito" factory-method="mock">
 		<constructor-arg value="edu.unc.lib.cdr.CdrEventProcessor" />
 	</bean>
 	
-	<camel:camelContext id="CdrServiceSolrUpdate">
-		<camel:package>edu.unc.lib.cdr.solrUpdate</camel:package>
+	<camel:camelContext id="CdrServiceCdrEvents">
+		<camel:package>edu.unc.lib.cdr.cdrEvents</camel:package>
 	</camel:camelContext>
 	
 </beans>
